Add retrieval and objection models with logging helpers
