+++ apps/legal_discovery/models_trial.py
@@ -66,27 +66,3 @@ class StrategySuggestion(db.Model):
     confidence = db.Column(db.Integer)
 
 
-class ObjectionEvent(db.Model):
-    __tablename__ = "objection_events"
-    id = db.Column(db.String, primary_key=True, default=uuid4)
-    session_id = db.Column(db.String, index=True)
-    segment_id = db.Column(db.String, index=True)
-    trace_id = db.Column(db.String(40), index=True)
-    ts = db.Column(db.DateTime, default=datetime.utcnow)
-    type = db.Column(db.String)
-    ground = db.Column(db.String)
-    confidence = db.Column(db.Integer)
-    extracted_phrase = db.Column(db.String)
-    suggested_cures = db.Column(db.JSON)
-    refs = db.Column(db.JSON)
-    path = db.Column(db.JSON)
-    action_taken = db.Column(db.String)
-    outcome = db.Column(db.String)
-
-
-class ObjectionResolution(db.Model):
-    __tablename__ = "objection_resolutions"
-    id = db.Column(db.String, primary_key=True, default=uuid4)
-    event_id = db.Column(db.String, db.ForeignKey("objection_events.id"), index=True)
-    chosen_cure = db.Column(db.String)
-    ts = db.Column(db.DateTime, default=datetime.utcnow)
