+++ tests/apps/test_objection_route.py
@@ -1,15 +1,11 @@
 from flask import Flask
 
 from apps.legal_discovery.extensions import socketio
-from apps.legal_discovery.database import db, RetrievalTrace
+from apps.legal_discovery.database import db
 from apps.legal_discovery.trial_assistant import bp as trial_bp
 from apps.legal_discovery.hippo_routes import objections_bp
-from apps.legal_discovery.models_trial import (
-    TrialSession,
-    ObjectionEvent,
-    ObjectionResolution,
-    TranscriptSegment,
-)
+from apps.legal_discovery.models import ObjectionEvent, ObjectionResolution, RetrievalTrace
+from apps.legal_discovery.models_trial import TrialSession, TranscriptSegment
 from apps.legal_discovery.trial_assistant.services.objection_engine import engine
 from apps.legal_discovery import hippo
