+++ tests/coded_tools/legal_discovery/test_objection_engine.py
@@ -1,16 +1,19 @@
 from types import SimpleNamespace
 
-from apps.legal_discovery.trial_assistant.services.objection_engine import ObjectionEngine, db
-
-
-def noop(*args, **kwargs):
-    return None
+from apps.legal_discovery.trial_assistant.services.objection_engine import ObjectionEngine
+from apps.legal_discovery.models import ObjectionEvent
 
 
 def test_counter_objection_detection(monkeypatch):
     engine = ObjectionEngine()
-    monkeypatch.setattr(db.session, "add", noop)
-    monkeypatch.setattr(db.session, "commit", noop)
+
+    def fake_log(**kwargs):
+        return ObjectionEvent(**kwargs)
+
+    monkeypatch.setattr(
+        "apps.legal_discovery.trial_assistant.services.objection_engine.log_objection_event",
+        fake_log,
+    )
     seg = SimpleNamespace(id="s1", text="That was an excited utterance")
     events = engine.analyze_segment("sess", seg)
     assert any(e.type == "counter" and e.ground == "hearsay" for e in events)
